Add files via upload
The rise of ML algorithms has necessitated a steep growth in the need for computational complexity, highlighting the von Neumann bottlenecks between AI computing hardware and data storage. The idea behind compute in-memory is simple: instead of having distinct compartments for memory and processing, the operations are 
performed in memory itself.